Squashed commit of the following:
commit c741343a22582fc50db991aa115fa40662d619fe
Merge: be6acaa 45f314c
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Thu Jul 12 23:10:15 2012 +0200

    Merge branch 'snuparek_macosx' into snuparek

commit be6acaab5e85c99bcfa6a8e4377449572c6447f7
Merge: 96c0683 0ab0707
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Thu Jul 12 23:09:59 2012 +0200

    Merge branch 'snuparek' of 192.168.0.3:/home/snupa3am/git/zlomekfs into snuparek

commit 45f314cda3beb00253306b24551f9b05299a86c3
Author: U-HP-F5FE4398DF84\2510p <snuparek@atlas.cz>
Date:   Thu Jul 12 14:07:23 2012 +0200

    Fix issues after Mac OS X port

    modified:   ../CMakeLists.txt

commit 2286294109cac50fef62c51100892b38f080afbd
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Thu Jul 12 14:05:52 2012 +0200

    Fix some issues after Mac OS X port

    modified:   zfsd/lib/zfs_dirent/zfs_dirent.c

commit 2b60d9c1866f9d1f9716142137d6e12040fefa7c
Author: admin1 <snuparek@atlas.cz>
Date:   Thu Jul 12 13:42:25 2012 +0200

    Ported zlomekFS to MacOS X 10.7

    modified:   CMakeLists.txt
    modified:   makeall.sh
    modified:   regression-testing/syplog/src/CMakeLists.txt
    modified:   regression-testing/syplog/src/log-constants.h
    deleted:    regression-testing/syplog/src/media/medium-api.c
    modified:   zfsd/CMakeLists.txt
    modified:   zfsd/configuration/zfs_config/zfs_config.c
    modified:   zfsd/control/dbus/dbus-provider.c
    modified:   zfsd/fs-entities/CMakeLists.txt
    modified:   zfsd/fs-entities/file/CMakeLists.txt
    modified:   zfsd/fs-entities/file/file.c
    modified:   zfsd/fs-entities/metadata/metadata.c
    modified:   zfsd/lib/CMakeLists.txt
    modified:   zfsd/lib/threading/pthread-wrapper.c
    modified:   zfsd/lib/threading/pthread-wrapper.h
    modified:   zfsd/lib/threading/thread.c
    new file:   zfsd/lib/zfs_dirent/CMakeLists.txt
    new file:   zfsd/lib/zfs_dirent/zfs_dirent.c
    new file:   zfsd/lib/zfs_dirent/zfs_dirent.h
    modified:   zfsd/lib/zfsio/zfsio_mkstemp.c
    modified:   zfsd/version/CMakeLists.txt
    modified:   zfsd/version/version.c

commit 0ab07077453d281dd709e5e4551471dc79b7bdad
Author: U-iMac\admin <snuparek@atlas.cz>
Date:   Tue Jul 3 17:49:48 2012 +0200

    Fix some errors

    modified:   conf/single/etc/zfsd/zfsd.conf
    modified:   zfsd/lib/CMakeLists.txt
    modified:   zfsd/lib/dokan_iface/tests/CMakeLists.txt
    modified:   zfsd/lib/dokan_iface/tests/POSIXapi_test/dir_tests.c
    modified:   zfsd/lib/dokan_iface/tests/POSIXapi_test/file_tests.c
    modified:   zfsd/lib/dokan_iface/tests/syscall_collector/CMakeLists.txt
    modified:   zfsd/lib/dokan_iface/tests/syscall_collector/syscall_collector.c
    modified:   zfsd/zfsd/zfsd.c

commit 6551682bcc53502b283e7a7f30c5e9f652bfe7c0
Merge: 05c6328 82c072c
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Mon Jul 2 14:04:40 2012 +0200

    Merge branch 'snuparek' of 192.168.0.3:/home/snupa3am/git/zlomekfs into snuparek

commit 05c6328183545f4fceb084bf050463b79ba92f59
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Wed Mar 14 20:45:29 2012 +0100

    Code cleanup

    modified:   lib/threading/thread.c
    modified:   version/version.c

commit 248c93bb57c106c23c289a4f5501ba9002db237b
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Wed Mar 14 20:43:38 2012 +0100

    Change protocol buffer namespace

    modified:   lib/protocol/protobuf/client.c
    modified:   lib/protocol/protobuf/server.c
    modified:   lib/protocol/protobuf/zfs.proto

commit ee1ccc1cd00b52384f7cab32822e459c0d95d0d5
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Wed Mar 14 18:58:05 2012 +0100

    Remove old unused kernel RPC call

    modified:   ../fuse_iface/kernel.h
    modified:   zfs-prot.def
    modified:   zfs-prot.h

commit a9d7e7ece725560cf2e15f609754a4837ee47415
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Fri Mar 2 15:42:28 2012 +0100

    Improve protocol buffers mockup

    modified:   client.c
    modified:   server.c
    modified:   zfs.proto

commit a4772ea54c904ea09e52d3a15bfa63cce9a942de
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Fri Mar 2 00:08:34 2012 +0100

    Write simple protocol buffers tutorial

    modified:   CMakeLists.txt
    new file:   client.c
    deleted:    main.c
    new file:   server.c
    modified:   zfs.proto

commit d450de7162927a006d8b1e010a13f08dc9e305f6
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Thu Mar 1 21:12:22 2012 +0100

    Fix wrong refactored code

    modified:   zfsd/lib/network/network.c
    modified:   zfsd/update/update.c

commit bff95d0c74d4d789815c154b215542388ac77411
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Sat Feb 18 00:00:01 2012 +0100

    Add tags to gitignore

    modified:   .gitignore

commit 888853e5b382739d6e4e93669a390e1e4c0d324c
Author: Ales Snuparek <snuparek@atlas.cz>
Date:   Wed Feb 15 01:17:08 2012 +0100

    Call create_this_node before reading local volume configuration

    modified:   .gitignore
    modified:   zfsd/configuration/config_reader/local_config.c
    modified:   zfsd/zfsd/zfsd.c

git-svn-id: https://shiva.ms.mff.cuni.cz/svn/zlomekfs/branches/snuparek@3280 b158dfce-25fb-0310-9a5b-f29f65ed45b5

+++ zfsd/fs-entities/metadata/metadata.c
@@ -2835,7 +2835,7 @@ void get_local_path_from_metadata(string * path, volume vol, zfs_fh * fh)
 			free(parent_path.str);
 
 			if (lstat(path->str, &st) != 0
-				|| st.st_dev != fh->dev || st.st_ino != fh->ino)
+				|| (uint32_t)st.st_dev !=  fh->dev || st.st_ino != fh->ino)
 			{
 				flush |= hardlink_list_delete_entry(hl, entry);
